krdma: remove send msg pool

+++ modules/include/rack_dvs.h
@@ -36,7 +36,7 @@ struct dvs_node {
 #define dvs_write(dvsr, addr, offset, size) \
     dvs_io(dvsr, addr, offset, size, WRITE)
 
-int dvs_io(struct dvs_region *dvsr, dma_addr_t addr, u64 offset, u32 size, int dir);
+int dvs_io(struct dvs_region *dvsr, dma_addr_t addr, u64 offset, u64 size, int dir);
 struct dvs_region *dvs_alloc_region(u64 size_mb, u64 slab_size_mb);
 void dvs_free_region(struct dvs_region *dvsr);
